--- pkg/devfile/adapters/common/interface.go
@@ -1,6 +1,0 @@
-package common
-
-// ComponentAdapter defines the functions that platform-specific adapters must implement
-type ComponentAdapter interface {
-	Push(parameters PushParameters) error
-}
