+++ pkg/devfile/adapters/kubernetes/component/interface.go
@@ -1,0 +1,8 @@
+package component
+
+import "github.com/redhat-developer/odo/pkg/devfile/adapters/common"
+
+// ComponentAdapter defines the functions that platform-specific adapters must implement
+type ComponentAdapter interface {
+	Push(parameters common.PushParameters) error
+}
